Extract integration test spec super class
